python/wg/requirements.txt
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
idna==2.6
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.4.0
